fix spiview and minor bug with initial search
